chore: getReactGridLayoutConfig & getResponsiveGridLayoutProps
