+++ packages/frontend/src/pages/Dashboard.tsx
@@ -1,7 +1,6 @@
 import {
     assertUnreachable,
     Dashboard as IDashboard,
-    DashboardTile,
     DashboardTileTypes,
     isDashboardChartTileType,
 } from '@lightdash/common';
@@ -51,30 +50,10 @@ import { useDashboardContext } from '../providers/DashboardProvider/useDashboard
 import { TrackSection } from '../providers/TrackingProvider/TrackSection';
 import '../styles/react-grid.css';
 import { SectionName } from '../types/Events';
-
-export const getReactGridLayoutConfig = (
-    tile: DashboardTile,
-    isEditMode = false,
-): Layout => ({
-    minH: 1,
-    minW: 6,
-    x: tile.x,
-    y: tile.y,
-    w: tile.w,
-    h: tile.h,
-    i: tile.uuid,
-    isDraggable: isEditMode,
-    isResizable: isEditMode,
-});
-
-export const getResponsiveGridLayoutProps = (enableAnimation = true) => ({
-    draggableCancel: '.non-draggable',
-    useCSSTransforms: enableAnimation,
-    measureBeforeMount: !enableAnimation,
-    breakpoints: { lg: 1200, md: 996, sm: 768 },
-    cols: { lg: 36, md: 30, sm: 18 },
-    rowHeight: 50,
-});
+import {
+    getReactGridLayoutConfig,
+    getResponsiveGridLayoutProps,
+} from '../utils/reactGridLayout';
 
 const ResponsiveGridLayout = WidthProvider(Responsive);
