+++ packages/frontend/src/utils/reactGridLayout.ts
@@ -1,0 +1,26 @@
+import { DashboardTile } from '@lightdash/common';
+import { Layout } from 'react-grid-layout';
+
+export const getReactGridLayoutConfig = (
+    tile: DashboardTile,
+    isEditMode = false,
+): Layout => ({
+    minH: 1,
+    minW: 6,
+    x: tile.x,
+    y: tile.y,
+    w: tile.w,
+    h: tile.h,
+    i: tile.uuid,
+    isDraggable: isEditMode,
+    isResizable: isEditMode,
+});
+
+export const getResponsiveGridLayoutProps = (enableAnimation = true) => ({
+    draggableCancel: '.non-draggable',
+    useCSSTransforms: enableAnimation,
+    measureBeforeMount: !enableAnimation,
+    breakpoints: { lg: 1200, md: 996, sm: 768 },
+    cols: { lg: 36, md: 30, sm: 18 },
+    rowHeight: 50,
+});
